Aggiunta funzione conferma time-sheet Aggiunto reminder via mail ogni ultimo del mese Aggiunta feature lista utenti per vedere chi non ha consegnato il time sheet Aggiunta lista utenti che non hanno compilato per un determinato mese in view Cambiato separatore breadcrumb

+++ src/main/java/eu/Rationence/pat/controller/RoleController.java
@@ -9,6 +9,7 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -37,8 +38,27 @@ public class RoleController {
     }
 
     @PostMapping("/roles")
-    public ResponseEntity<String> addRole(@Valid Role role) {
+    public ResponseEntity<String> addRole(@Valid Role role,
+                                          BindingResult result) {
         try {
+            if(result.hasErrors())
+                return AdviceController.responseBadRequest(result.getAllErrors().toString());
+            roleService.save(role);
+            return AdviceController.responseOk("Role '" + role.getRoleName() + "' saved.");
+        } catch (Exception e) {
+            return AdviceController.responseServerError(e.getMessage());
+        }
+
+    }
+
+    @PutMapping("/roles")
+    public ResponseEntity<String> updateRole(@Valid Role role,
+                                          BindingResult result) {
+        try {
+            if(result.hasErrors())
+                return AdviceController.responseBadRequest(result.getAllErrors().toString());
+            if(roleService.findRole(role.getRoleName()) == null)
+                return AdviceController.responseNotFound("Cannot update Role " + role.getRoleName() + "(not found)");
             roleService.save(role);
             return AdviceController.responseOk("Role '" + role.getRoleName() + "' saved.");
         } catch (Exception e) {
@@ -48,10 +68,13 @@ public class RoleController {
     }
 
     @DeleteMapping("/roles")
-    public ResponseEntity<String> deleteRole(@Valid Role role) {
-        if (role.getRoleName().equals("ADMIN") || role.getRoleName().equals("USER"))
-            return AdviceController.responseForbidden("Can't delete default ADMIN or USER role");
+    public ResponseEntity<String> deleteRole(@Valid Role role,
+                                             BindingResult result) {
         try {
+            if (role.getRoleName().equals("ADMIN") || role.getRoleName().equals("USER"))
+                return AdviceController.responseForbidden("Can't delete default ADMIN or USER role");
+            if(result.hasErrors())
+                return AdviceController.responseBadRequest(result.getAllErrors().toString());
             roleService.deleteRole(role.getRoleName());
             return AdviceController.responseOk("Role '" + role.getRoleName() + "' deleted.");
         }
